# Add ability to disable trace event var plugging