--- a/topdown/trace_test.go
+++ b/topdown/trace_test.go
@@ -1010,3 +1010,16 @@ func TestShortTraceFileNames(t *testing.T) {
 		})
 	}
 }
+
+func TestBufferTracerTraceConfig(t *testing.T) {
+	ct := CustomTracer(NewBufferTracer())
+	conf := ct.Config()
+
+	expected := TraceConfig{
+		PlugLocalVars: true,
+	}
+
+	if !reflect.DeepEqual(expected, conf) {
+		t.Fatalf("Expected config: %+v, got %+v", expected, conf)
+	}
+}