This PR makes a few changes:

* Introduces a new `topdown.CustomTracer` interface which builds on top of the `topdown.Tracer`, but allows for Tracer implementations to supply their configuration requirements for evaluation time.
* Changes the `topdown.Query` and `topdown.Eval` to check if any tracers require variable metadata (via their configuration, as applicable). If none do then that step is skipped on each trace event.
* Updates the profiler and coverage tracers to take advantage of this, and reducing their overhead pretty significantly (Fixing #2245). Results from my dev machine before and after for coverage tracing on a query:

```
name                                                                old time/op    new time/op    delta
BenchmarkCoverBigLocalVar/1Vars1Iterations-8        41.0µs ± 3%    20.8µs ± 4%  -49.30%  (p=0.000 n=8+9)
BenchmarkCoverBigLocalVar/10Vars1Iterations-8       55.8µs ± 2%    23.7µs ±11%  -57.44%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/1Vars100Iterations-8      2.91ms ± 2%    0.38ms ± 7%  -87.07%  (p=0.000 n=9+10)
BenchmarkCoverBigLocalVar/10Vars100Iterations-8     35.4ms ± 1%     2.7ms ± 0%  -92.51%  (p=0.000 n=9+8)
BenchmarkCoverBigLocalVar/1Vars1000Iterations-8      156ms ± 0%       4ms ± 1%  -97.74%  (p=0.000 n=8+9)
BenchmarkCoverBigLocalVar/10Vars1000Iterations-8     1.17s ± 2%     0.03s ± 0%  -97.70%  (p=0.000 n=10+10)

name                                                                old alloc/op   new alloc/op   delta
BenchmarkCoverBigLocalVar/1Vars1Iterations-8        22.4kB ± 0%     9.9kB ± 0%  -55.68%  (p=0.002 n=8+10)
BenchmarkCoverBigLocalVar/10Vars1Iterations-8       25.9kB ± 0%    10.7kB ± 0%  -58.87%  (p=0.000 n=10+8)
BenchmarkCoverBigLocalVar/1Vars100Iterations-8      1.45MB ± 0%    0.18MB ± 0%  -87.30%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/10Vars100Iterations-8     17.7MB ± 0%     1.2MB ± 0%  -93.32%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/1Vars1000Iterations-8     65.4MB ± 0%     1.8MB ± 0%  -97.31%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/10Vars1000Iterations-8     498MB ± 0%      12MB ± 0%  -97.56%  (p=0.000 n=9+10)

name                                                                old allocs/op  new allocs/op  delta
BenchmarkCoverBigLocalVar/1Vars1Iterations-8           355 ± 0%       171 ± 0%  -51.83%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/10Vars1Iterations-8          401 ± 0%       185 ± 0%  -53.87%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/1Vars100Iterations-8       13.9k ± 0%      4.6k ± 0%  -66.81%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/10Vars100Iterations-8       133k ± 0%       30k ± 0%  -77.75%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/1Vars1000Iterations-8       138k ± 0%       46k ± 0%  -66.70%  (p=0.000 n=10+10)
BenchmarkCoverBigLocalVar/10Vars1000Iterations-8     1.38M ± 0%     0.31M ± 0%  -77.87%  (p=0.000 n=8+10)
```

The buffer tracer was left with its original behavior, and any existing tracers will continue to work the same way as before. Implementors must opt-in by adding the new required functions to meet the `topdown.CustomTracer` interface.